tentacle/requirements.txt
flask
gunicorn
ruamel-yaml
requests
